Refactor stats util
